pharmaversesdtm: SDTM Test Data for the 'Pharmaverse' Family of Packages

A set of Study Data Tabulation Model (SDTM) datasets from the Clinical Data Interchange Standards Consortium (CDISC) pilot project used for testing and developing Analysis Data Model (ADaM) datasets inside the pharmaverse family of packages. SDTM dataset specifications are described in the CDISC SDTM implementation guide, accessible by creating a free account on https://www.cdisc.org/.

See also

Author

Maintainer: Lina Patil lina.patil@cytel.com

Authors:

Stefan Bundfuss (ORCID)

Fanny Gautier fanny.gautier@cytel.com (ORCID)

Edoardo Mancini edoardo.mancini@roche.com (ORCID)

Tomoyuki Namai

Vinh Nguyen

Vladyslav Shuliar (ORCID)

Other contributors:

Cytel Inc. [copyright holder, funder]

F. Hoffmann-La Roche AG [copyright holder, funder]

GlaxoSmithKline LLC [copyright holder, funder]