Questionnaire Dataset-metabolic

A SDTM QS dataset containing COEQ data (Control of Eating Questionnaire) for metabolic studies. Note that University of Leeds are the copyright holders of the Control of Eating Questionnaire (CoEQ) and the test data included within pharmaversesdtm as well as the ADCOEQ code are for not-for-profit use only within admiralmetabolic and pharmaverse-related examples/documentation. Any persons or companies wanting to use the CoEQ should request a license to do so from the following link.

Usage

qs_metabolic

Format

A data frame with 18 columns:

STUDYID

Study Identifier

USUBJID

Unique Subject Identifier

DOMAIN

Domain Abbreviation

VISIT

Visit Name

VISITNUM

Visit Number

VISITDY

Planned Study Day of Visit

QSBLFL

Baseline Flag

QSDTC

Date/Time of Finding

QSDY

Study Day of Finding

QSCAT

Category for Questionnaire

QSTEST

Questionnaire Test Name

QSTESTCD

Questionnaire Test Short Name

QSORRES

Result or Finding in Original Units

QSORRESU

Original Units

QSSTRESC

Character Result/Finding in Std Format

QSSTRESN

Numeric Result/Finding in Standard Units

QSSTRESU

Standard Units

QSSEQ

Sequence Number

Source

Constructed by admiralmetabolic developers

Details

Questionnaire Dataset-metabolic

A SDTM QS dataset containing COEQ data (Control of Eating Questionnaire) for metabolic studies. Note that University of Leeds are the copyright holders of the Control of Eating Questionnaire (CoEQ) and the test data included within pharmaversesdtm as well as the ADCOEQ code are for not-for-profit use only within admiralmetabolic and pharmaverse-related examples/documentation. Any persons or companies wanting to use the CoEQ should request a license to do so from the following link.

Contains a set of 21 unique Test Short Names and Test Names:

QSTESTCD	QSTEST

COEQ01	How hungry have you felt?

COEQ02	How full have you felt?

COEQ03	How strong was your desire to eat sweet foods?

COEQ04	How strong was your desire to eat savoury foods?

COEQ05	How happy have you felt?

COEQ06	How anxious have you felt?

COEQ07	How alert have you felt?

COEQ08	How contented have you felt?

COEQ09	During the last 7 days how often have you had food cravings?

COEQ10	How strong have any food cravings been?

COEQ11	How difficult has it been to resist any food cravings?

COEQ12	How often have you eaten in response to food cravings?

COEQ13	Chocolate or chocolate flavoured foods

COEQ14	Other sweet foods (cakes, pastries, biscuits, etc)

COEQ15	Fruit or fruit juice

COEQ16	Dairy foods (cheese, yoghurts, milk, etc)

COEQ17	Starchy foods (bread, rice, pasta, etc)

COEQ18	Savoury foods (french fries, crisps, burgers, pizza, etc)

COEQ19	Generally, how difficult has it been to control your eating?

COEQ20	Which one food makes it most difficult for you to control eating?

COEQ21	How difficult has it been to resist eating this food during the last 7 days?